change doubles to floats

+++ ReadADC/ReadADC.cpp
@@ -86,40 +86,40 @@ void ADCSensor::reset(){
 
 void ADCSensor::calculate(){
 
-  double mV = 0.00; // initialize a temp variable only once to save processing power
+  float mV = 0.00; // initialize a temp variable only once to save processing power
 
   // calculate the average values
-  double sensor_read_avg = (this->running_read_total_) / (this->running_read_count_);
+  float sensor_read_avg = (this->running_read_total_) / (this->running_read_count_);
   mV = (sensor_read_avg / 4095.0) * 5000; // read 0 = 0V, 4095 = 5V. 5000mV = 5V.
   this->actual_avg_ = (mV - this->zero_mV_) / this->mV_per_sensor_unit_; // calculates the sensor value
 
   // calculate the min values
-  mV = (static_cast<double>(this->read_min_) / 4095.0) * 5000;
+  mV = (static_cast<float>(this->read_min_) / 4095.0) * 5000;
   this->actual_min_ = (mV - this->zero_mV_) / this->mV_per_sensor_unit_;
 
   // calculate the max values
-  mV = (static_cast<double>(this->read_max_) / 4095.0) * 5000;
+  mV = (static_cast<float>(this->read_max_) / 4095.0) * 5000;
   this->actual_max_ = (mV - this->zero_mV_) / this->mV_per_sensor_unit_;
 
   this->reset();
   this->values_updated = true;
 }
 
-double ADCSensor::avg(){
+float ADCSensor::avg(){
   if (!values_updated){ // if the values are not the most current, update them before returning the average!
     this->calculate();
   }
   return this->actual_avg_;
 }
 
-double ADCSensor::min(){
+float ADCSensor::min(){
   if (!values_updated){ // if the values are not the most current, update them before returning the average!
     this->calculate();
   }
   return this->actual_min_;
 }
 
-double ADCSensor::max(){
+float ADCSensor::max(){
   if (!values_updated){ // if the values are not the most current, update them before returning the average!
     this->calculate();
   }
